Rename files component
