+++ src/components/Footer/index.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 
 import style from './Footer.module.scss';
 
-function Footer() {
+function Index() {
   return <div className={style.footer}>Footer</div>;
 }
 
-export default Footer;
+export default Index;
